+++ CHANGELOG.md
@@ -3,6 +3,18 @@ Orion Core
 
 This is the changelog for ```orionjs:core```.
 
+### 0.7.5
+
+- Choose fields for custom entity permissions.
+
+### 0.7.4
+
+- Entity index table doesn't listen to click if it is <a> or <button>
+
+### 0.7.3
+
+- In /login you can set ?redirect=xxx.
+
 ### 0.7.2
 
 - "Created by" on entities is no more required (to insert values on server).